requirements.txt
google-generativeai==0.3.2
streamlit==1.28.0
opencv-python==4.8.1.78
Pillow==10.0.0
python-dotenv==1.0.0
speech_recognition==3.10.0
pyttsx3==2.90
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
